include: drivers: uart: remove incorrect comments about CDC ACM UART
The CDC ACM implementation provides a virtual UART interface
that is used by various samples and subsystems in a way that is
compatible with the real UART interface.

Commit cc1b2c70cc8b
("uart: doc: Add special case for virtual UART")
added a questionable comment to uart_fifo_fill() API
description in an attempt to fix issue described in
https://github.com/zephyrproject-rtos/zephyr/issues/11455

However, this did not fix the problem because the API is not
designed to be used in this way.  Finally commit 0e57e4fb78e0
("samples: usb: cdc_acm: Update CDC ACM echo sample")
revised the sample to use the API in the correct way.

Remove incorrect comment in uart_fifo_fill(). For compatibility
reasons and due to the design of the API, such differences must not
exist.

(cherry picked from commit 4d3611e95c702bd209355dd03414a73929a753c9)

Original-Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>
GitOrigin-RevId: 4d3611e95c702bd209355dd03414a73929a753c9
Change-Id: If9cb043aa16b587b1e9bbd060f6c8ff4303b53e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4667073
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>