kconfig: Set `BOOTLOADER_SRAM_SIZE` default to 0
The current default value for `BOOTLOADER_SRAM_SIZE` is set to an
arbitrary-seeming value of `16`, which feels like a random magic
number.

Given that adding an offset in SRAM for a bootloader should always be
a conscious choice for a specific platform, this PR sets the default
to `0`. An appropriate value can be set at the SoC or board level,
overriding this default.

(cherry picked from commit f68cee94057a7a9e75ccda78aee952ded51f8156)

Original-Signed-off-by: Kevin Townsend <kevin.townsend@linaro.org>
GitOrigin-RevId: f68cee94057a7a9e75ccda78aee952ded51f8156
Change-Id: I7ab1c99d3130e3ea71a8ca910394ea58582fc8dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4689755
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>