drivers: led: pca9633: disable allcall
The PCA9633 i2c LED controller offers an All Call address in its nominal
operation, allowing simultaneous communication with all instances present
on the same i2c bus. The default address is 0x70. While this functionality
is convenient, it is possible that the board uses another i2c component
that also uses this address (for example, the shtcx). In such cases, the
address conflict prevents the proper functioning of the system.

The idea is to add a "disable-allcall" property to the device tree. If this
option is present, the initialization of the PCA9633 forces the bit 0
(ALLCALL) to be set to false, thereby disabling this function. It is
necessary to add this property to all PCA9633 devices on the bus to free up
the address 0x70.

(cherry picked from commit 13d74677bac1d2ecf3be356e0ed6ce626e199921)

Original-Signed-off-by: Steve Jacot-Guillarmod <steve@piziwate.net>
GitOrigin-RevId: 13d74677bac1d2ecf3be356e0ed6ce626e199921
Change-Id: I9a364ab78605d39d29c812c17882b2ee3a35f934
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4920696
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>