net: tcp: Remove trigger of send_data_timer when window full
Likely this trigger of the send_data_timer was an alternative for the
function that has been filled in by the ZWP transmission.
At the moment this timer has the potential to cause spurious
retransmissions that can degrade the throughput of the network stack.
Second to that it can accelerate the retransmission process, quickly
running to the number of retransmissions, causing a connection failure.

(cherry picked from commit 1f3cb08fdc1481202d0f8e708e28d3239520fc91)

Original-Signed-off-by: Sjors Hettinga <s.a.hettinga@gmail.com>
GitOrigin-RevId: 1f3cb08fdc1481202d0f8e708e28d3239520fc91
Change-Id: I68bda94a7558a2fe04b5e64cc102b50bdec98365
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4738212
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>