Bluetooth: audio: ascs: Fix ASE state transition due to CIS link loss
This fixes ASE state transition to QoS Configured state that shall be
done autononously when CIS has been terminated due to link loss.

As per ASCS v1.0:
"If the server detects link loss of a CIS for an ASE in the Streaming
state or the Disabling state, the server shall immediately transition
that ASE to the QoS Configured state. Link loss of a CIS for an ASE in
any state other than Streaming or Disabling shall not cause a
transition of the ASE state machine."

(cherry picked from commit 8dcec59115bd62951ff364f70aa450458d367b68)

Original-Fixes: #60669
Original-Signed-off-by: Mariusz Skamra <mariusz.skamra@codecoup.pl>
GitOrigin-RevId: 8dcec59115bd62951ff364f70aa450458d367b68
Change-Id: I12253e3efa781ad3d1b129de2ac2ada581264942
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4789627
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>