native_posix: IRQ handling: dont swap if kernel is uninitialized
After an interrupt, do not attempt to swap
if the kernel is not yet initialized.

Otherwise, if an interrupt is raised with interrupts
unlocked while the kernel is not yet fully initialized,
a swap would lead to a crash.
(This could happen if a PREKERNEK1/2 driver enables interrupts
and an interrupt fires before the kernel has swapped to
main)

(cherry picked from commit dcd6ee518867de3d2ff02f5b9dea084feef17821)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: dcd6ee518867de3d2ff02f5b9dea084feef17821
Change-Id: I8420b48e907b7b3be5837e691582c2348e578f19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4764008
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>