drivers: spi: spi_nrfx_spi: Add CPOL handling on SCK pin
Pin state after SPI deinitialization is based on pinctrl configuration.
On the other hand, CPOL is set during runtime. When the SPI instance
is disabled GPIO takes control over SCK and drives it to state set
by pinctrl driver. This might causes an invalid SCK state
when the transaction is configured with CPOL (Clock Polarity).

To address this issue, a patch was introduced to the SPI driver.
Now, when a SPI instance is configured with CPOL,
the driver is setting in the runtime the correct state of the SCK pin.

(cherry picked from commit 9de69aabfddc60c73f8897e38463ef2479201375)

Original-Signed-off-by: Adam Wojasinski <adam.wojasinski@nordicsemi.no>
GitOrigin-RevId: 9de69aabfddc60c73f8897e38463ef2479201375
Change-Id: Id128a4a53ea6ef774feda6509a3fb2762e2e4a72
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4769256
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>