scripts: coccinelle: Remove strtok_r symbol from checklist
strtok_r is part of GNU99, but default Zephyr is built with C99, so,
the GNU headers are not pulled in, even if we add "--std=gnu99" it
won't override the C99 standard.

As a workaround for build we redefine the prototype, but this causes a
coding guideline warning, so, for now disable this warning
for strtok_r.

(cherry picked from commit 0399213cd5506ab9d726eeac1e51a5b864eb38b4)

Original-Signed-off-by: Sachin D Kulkarni <sachin.kulkarni@nordicsemi.no>
GitOrigin-RevId: 0399213cd5506ab9d726eeac1e51a5b864eb38b4
Change-Id: I209dc426acc670ac52bef616cee465684f44ccbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4808567
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>