FROMPULL: scripts: zephyr_module: Tolerate symlinks
Currently, this script will blow up when given a symlink for a path (e.g.,
to a kconfig file).

This is annoying, as when wrapping the build system in a hermetic build
system like Bazel, Bazel likes to limit the input file set by creating
symlinks.

Resolve the path, if it is a file once resolved, then it's OK.

https://github.com/zephyrproject-rtos/zephyr/pull/61352

BUG=b:274648575
TEST=CQ

Change-Id: Ic002946737bb41025e09b1b5dec33fb29dab7987
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4771184
Reviewed-by: Aaron Massey <aaronmassey@google.com>